fix: filter category
