+++ src/components/RenderList/index.jsx
@@ -33,8 +33,8 @@ const RenderList = ({
   register,
   Search,
   filteredRender,
-  events,
-  setFilteredRender,
+  eventsIJoined,
+  eventsICreated
 }) => {
   const { user } = useUser();
   const UserID = localStorage.getItem("UserID");
@@ -42,51 +42,7 @@ const RenderList = ({
     e.preventDefault();
     Search();
   };
-  function eventsICreated(idUser) {
-    const eventsCreated = events.filter((event) => event.userId == idUser);
-    if (eventsCreated.length < 1) {
-      toast.error("Você ainda não criou evento!", {
-        position: "top-right",
-        autoClose: 5000,
-        hideProgressBar: false,
-        closeOnClick: true,
-        pauseOnHover: true,
-        draggable: true,
-        progress: undefined,
-      });
-    } else {
-      setFilteredRender(eventsCreated);
-      return eventsCreated;
-    }
-  }
 
-  function eventsIJoined(idUser) {
-    const arrEvents = [];
-    const eventsJoined = events.map((event) =>
-      event.guests.filter((personId) => {
-        return personId.id == idUser && event;
-      })
-    );
-    for (let i = 0; i < eventsJoined.length; i++) {
-      if (eventsJoined[i].length > 0) {
-        arrEvents.push(events[i]);
-      }
-    }
-    if (arrEvents.length < 1) {
-      toast.error("Você não entrou em evento!", {
-        position: "top-right",
-        autoClose: 5000,
-        hideProgressBar: false,
-        closeOnClick: true,
-        pauseOnHover: true,
-        draggable: true,
-        progress: undefined,
-      });
-    } else {
-      setFilteredRender(arrEvents);
-      return arrEvents;
-    }
-  }
 
   return (
     <>
